requirements.txt
absl-py==2.0.0
addict==2.4.0
aioboto3==11.3.0
aiobotocore==2.6.0
aiofiles==23.2.1
aiohttp==3.8.5
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.12.0
altair==5.1.2
annotated-types==0.5.0
anyio==3.7.1
arrow==1.3.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.10.1
blessed==1.20.0
blinker==1.6.3
boto3==1.28.17
botocore==1.31.17
bottle==0.12.25
Brotli @ file:///work/ci_py311/brotli-split_1676830125088/work
cachetools==5.3.1
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpickle==2.2.1
cmaes==0.10.0
colorama==0.4.6
colorlog==6.7.0
contourpy==1.1.1
croniter==1.4.1
cryptography @ file:///croot/cryptography_1702070282333/work
cycler==0.12.0
databricks-cli==0.18.0
dateutils==0.6.12
deepdiff==6.6.0
deeplake==3.8.0
dill==0.3.7
docker==6.1.3
einops==0.7.0
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.103.2
ffmpy==0.3.1
filelock @ file:///croot/filelock_1700591183607/work
flake8 @ file:///work/ci_py311/flake8_1677709235205/work
Flask==2.3.3
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
gmpy2 @ file:///work/ci_py311/gmpy2_1676839849213/work
google-auth==2.23.2
google-auth-oauthlib==1.0.0
gradio_client==0.7.0
greenlet==3.0.0
grpcio==1.59.0
gunicorn==21.2.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
hub==3.0.1
huggingface-hub==0.19.4
humbug==0.3.2
idna @ file:///work/ci_py311/idna_1676822698822/work
imageio==2.31.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2 @ file:///work/ci_py311/jinja2_1676823587943/work
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
kmeans-pytorch==0.3
kornia==0.7.0
lazy_loader==0.3
libdeeplake==0.0.83
lightning==2.1.2
lightning-cloud==0.5.39
lightning-utilities==0.9.0
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe @ file:///work/ci_py311/markupsafe_1676823507015/work
matplotlib==3.8.0
mccabe @ file:///opt/conda/conda-bld/mccabe_1644221741721/work
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mlflow==2.7.1
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx @ file:///croot/networkx_1690561992265/work
numcodecs==0.12.0
numpy @ file:///croot/numpy_and_numpy_base_1701295038894/work/dist/numpy-1.26.2-cp311-cp311-linux_x86_64.whl#sha256=14973c61f3a96f31efa18449e9a0ba592baf1d942a48d1f461f877ee439e848f
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
optuna==3.5.0
optuna-dashboard==0.12.0
ordered-set==4.1.0
orjson==3.9.10
packaging==23.2
pandas==2.1.1
pathos==0.3.1
pathspec==0.11.2
pep8==1.7.1
Pillow @ file:///croot/pillow_1696580024257/work
platformdirs==3.11.0
pox==0.3.3
ppft==1.7.6.7
prettytable==3.9.0
protobuf==4.24.4
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle @ file:///work/ci_py311/pycodestyle_1677708783451/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.9.1
pydantic_core==2.14.5
pydub==0.25.1
pyflakes @ file:///work/ci_py311/pyflakes_1677708650266/work
pygame==2.5.2
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9.post0
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
ray==2.8.1
readchar==4.0.5
referencing==0.30.2
regex==2023.10.3
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.4
rsa==4.9
s3transfer==0.6.2
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
starlette==0.27.0
starsessions==1.3.0
sympy @ file:///croot/sympy_1701397643339/work
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.1
torch-tb-profiler==0.4.3
torchaudio==2.1.0
torchdata==0.7.1
torchmetrics==1.2.0
torchsummary==1.5.1
torchtext==0.16.1
torchvision==0.16.1
tqdm==4.66.1
traitlets==5.11.2
transformers==4.35.2
triton==2.1.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1698257533958/work
uvicorn==0.23.2
wcwidth==0.2.8
websocket-client==1.6.3
websockets==11.0.3
Werkzeug==3.0.0
wrapt==1.15.0
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
